test: 테스트 모킹 클래스 외부로 분리
